11/04 - Managers Progress
ADD ProjectionGemsManager - Handles the usage, refund and increase of projection gems.
ADD ProjectionManager - Handles some events related to object projection (FailProjection, ProjectionSuccess), also checks if an object can be projected (communicates with ProjectionGemsManager for this purpose) --- It has to be responsable of the selection of the object to project(Not Implemented Yet)
ADD ProjectionPlatform Methods to handle object projection, also includes some events related to object projection (this events will be listened by some diegetic interface, maybe as a world space canvas inside the ProjectionPlatform gameObject).